- Rename HostWhitelistHandler to HostWhitelist - Add more unit test cases to HostWhitelist
